Minor refactors and doc comment tweaks
